isolates Group model tests
